backend/src/requirements.txt
python-dotenv
openai
requests
pydantic
Pillow
reportlab
rich
tenacity
fastapi
uvicorn
pydantic-settings
httpx

